// apps/server/src/db/schema/auth.d.ts
export declare const user: import("drizzle-orm/sqlite-core").SQLiteTableWithColumns<{
	name: "user";
	schema: undefined;
	columns: {
		id: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "id";
				tableName: "user";
				dataType: "string";
				columnType: "SQLiteText";
				data: string;
				driverParam: string;
				notNull: true;
				hasDefault: false;
				isPrimaryKey: true;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: [string, ...string[]];
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{
				length: number | undefined;
			}
		>;
		name: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "name";
				tableName: "user";
				dataType: "string";
				columnType: "SQLiteText";
				data: string;
				driverParam: string;
				notNull: true;
				hasDefault: false;
				isPrimaryKey: false;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: [string, ...string[]];
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{
				length: number | undefined;
			}
		>;
		email: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "email";
				tableName: "user";
				dataType: "string";
				columnType: "SQLiteText";
				data: string;
				driverParam: string;
				notNull: true;
				hasDefault: false;
				isPrimaryKey: false;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: [string, ...string[]];
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{
				length: number | undefined;
			}
		>;
		emailVerified: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "email_verified";
				tableName: "user";
				dataType: "boolean";
				columnType: "SQLiteBoolean";
				data: boolean;
				driverParam: number;
				notNull: true;
				hasDefault: false;
				isPrimaryKey: false;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: undefined;
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{}
		>;
		image: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "image";
				tableName: "user";
				dataType: "string";
				columnType: "SQLiteText";
				data: string;
				driverParam: string;
				notNull: false;
				hasDefault: false;
				isPrimaryKey: false;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: [string, ...string[]];
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{
				length: number | undefined;
			}
		>;
		role: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "role";
				tableName: "user";
				dataType: "string";
				columnType: "SQLiteText";
				data: string;
				driverParam: string;
				notNull: true;
				hasDefault: true;
				isPrimaryKey: false;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: [string, ...string[]];
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{
				length: number | undefined;
			}
		>;
		createdAt: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "created_at";
				tableName: "user";
				dataType: "date";
				columnType: "SQLiteTimestamp";
				data: Date;
				driverParam: number;
				notNull: true;
				hasDefault: false;
				isPrimaryKey: false;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: undefined;
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{}
		>;
		updatedAt: import("drizzle-orm/sqlite-core").SQLiteColumn<
			{
				name: "updated_at";
				tableName: "user";
				dataType: "date";
				columnType: "SQLiteTimestamp";
				data: Date;
				driverParam: number;
				notNull: true;
				hasDefault: false;
				isPrimaryKey: false;
				isAutoincrement: false;
				hasRuntimeDefault: false;
				enumValues: undefined;
				baseColumn: never;
				identity: undefined;
				generated: undefined;
			},
			{},
			{}
		>;
	};
	dialect: "sqlite";
}>;